requirements.txt
fastapi
redis
types-redis
uvicorn
sqlalchemy
psycopg2
python-dotenv
pydantic
sqlmodel
gunicorn
pyjwt 
passlib